+++ src/main/java/com/navi/stockexchange/models/Transaction.java
@@ -4,12 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
+import java.math.BigDecimal;
+
 @Data
 @EqualsAndHashCode
 @AllArgsConstructor
 public class Transaction {
     private String buyerOrderId;
-    private Float sellPrice;
+    private BigDecimal sellPrice;
     private int quantity;
     private String sellerOrderId;
 }
